feat: MovieInformation 내 오타로 인한 디코드 버그 수정, DailyBoxOfficeURL의 init 수정, Requestable 내 프로퍼티 수정
