+++ BoxOffice/Network/URL/Requestable.swift
@@ -4,9 +4,11 @@
 //
 //  Created by 리지, kokkilE on 2023/03/21.
 //
+
 import Foundation
+
 protocol Requestable {
     var url: URL? { get }
-//    var baseURL: String { get }
-//    var key: String { get }
+    var urlComponents: URLComponents? { get set }
+    var key: URLQueryItem { get }
 }
